추가 주석, 수정 getPerformanceDto 메서드
